+++ amazon-chime-sdk/src/test/java/com/amazonaws/services/chime/sdk/meetings/analytics/DefaultEventAnalyticsControllerTest.kt
@@ -102,7 +102,6 @@ class DefaultEventAnalyticsControllerTest {
         every { mockAppStateMonitor.appState } returns AppState.ACTIVE
         every { mockAppStateMonitor.getBatteryLevel() } returns 0.75f
         every { mockAppStateMonitor.getBatteryState() } returns BatteryState.CHARGING
-        every { mockAppStateMonitor.isBatterySaverOn() } returns true
 
         Dispatchers.setMain(testDispatcher)
     }
@@ -256,7 +255,6 @@ class DefaultEventAnalyticsControllerTest {
         assertEquals("Active", eventAttributes[EventAttributeName.appState.name])
         assertEquals(0.75f, eventAttributes[EventAttributeName.batteryLevel.name])
         assertEquals("Charging", eventAttributes[EventAttributeName.batteryState.name])
-        assertEquals(true.toString(), eventAttributes[EventAttributeName.lowPowerModeEnabled.name])
     }
 
     @Test
@@ -275,7 +273,6 @@ class DefaultEventAnalyticsControllerTest {
         assertEquals("Active", eventAttributes[EventAttributeName.appState.name])
         assertFalse(eventAttributes.containsKey(EventAttributeName.batteryLevel.name))
         assertEquals("Charging", eventAttributes[EventAttributeName.batteryState.name])
-        assertEquals(true.toString(), eventAttributes[EventAttributeName.lowPowerModeEnabled.name])
     }
 
     @Test
@@ -291,11 +288,10 @@ class DefaultEventAnalyticsControllerTest {
         assertEquals("Active", eventAttributes[EventAttributeName.appState.name])
         assertEquals(0.75f, eventAttributes[EventAttributeName.batteryLevel.name])
         assertEquals("Charging", eventAttributes[EventAttributeName.batteryState.name])
-        assertEquals(true.toString(), eventAttributes[EventAttributeName.lowPowerModeEnabled.name])
     }
 
     @Test
-    fun `pushHistory should include appState, batteryState, and lowPowerModeEnabled but not batteryLevel when battery level is null`() {
+    fun `pushHistory should include appState and batteryState but not batteryLevel when battery level is null`() {
         // Mock battery level as null
         every { mockAppStateMonitor.getBatteryLevel() } returns null
 
@@ -310,6 +306,5 @@ class DefaultEventAnalyticsControllerTest {
         assertEquals("Active", eventAttributes[EventAttributeName.appState.name])
         assertFalse(eventAttributes.containsKey(EventAttributeName.batteryLevel.name))
         assertEquals("Charging", eventAttributes[EventAttributeName.batteryState.name])
-        assertEquals(true.toString(), eventAttributes[EventAttributeName.lowPowerModeEnabled.name])
     }
 }